chore: move function to service instead of data access service
